Clear up training script
