+++ conf/config.yaml
@@ -1,17 +1,15 @@
 # Default configurations
 model:
-  architecture: alexnet
-  weights: ???
+  encoder: alexnet
+  weights: None
   input_size: (17, 17)
   output_size: (272, 272)
 data:
-  dataset: ImageNet
+  dataset: ImageNet2015
   dir: ./data/ILSVRC
-  train: ???
-  test: ???
-  target_size: (127, 127)
-  search_size: (255, 255)
+  exemplar_size: (127, 127)
+  target_size: (255, 255)
 hparams:
-  batch_size: 8
   max_epochs: 50
-  lr: (1e-2, 1e-5)
+  batch_size: 8
+  learning_rate: (1e-2, 1e-5)
