+++ environment.yml
@@ -4,81 +4,7 @@ channels:
   - conda-forge
   - defaults
 dependencies:
-  - defaults/osx-64::blas==1.0=mkl
-  - conda-forge/osx-64::bzip2==1.0.8=h0d85af4_4
-  - ca-certificates
-  - conda-forge/osx-64::giflib==5.2.1=hbcb3906_2
-  - conda-forge/osx-64::jbig==2.1=h0d85af4_2003
-  - conda-forge/osx-64::jpeg==9e=h0d85af4_0
-  - conda-forge/osx-64::lame==3.100=h35c211d_1001
-  - conda-forge/osx-64::libbrotlicommon==1.0.9=h0d85af4_6
-  - conda-forge/osx-64::libcxx==12.0.1=habf9029_1
-  - conda-forge/osx-64::libdeflate==1.10=h0d85af4_0
-  - conda-forge/osx-64::libffi==3.4.2=h0d85af4_5
-  - conda-forge/osx-64::libiconv==1.16=haf1e3a3_0
-  - conda-forge/osx-64::libuv==1.43.0=h0d85af4_0
-  - conda-forge/osx-64::libwebp-base==1.2.2=h0d85af4_1
-  - conda-forge/osx-64::libzlib==1.2.11=h9173be1_1013
-  - conda-forge/osx-64::llvm-openmp==13.0.1=hda6cdc1_0
-  - conda-forge/osx-64::ncurses==6.3=he49afe7_0
-  - conda-forge/osx-64::nettle==3.6=hedd7734_0
-  - conda-forge/osx-64::pthread-stubs==0.4=hc929b4f_1001
-  - conda-forge/noarch::tzdata==2021e=he74cb21_0
-  - conda-forge/osx-64::xorg-libxau==1.0.9=h35c211d_0
-  - conda-forge/osx-64::xorg-libxdmcp==1.1.3=h35c211d_0
-  - conda-forge/osx-64::xz==5.2.5=haf1e3a3_1
-  - conda-forge/osx-64::gettext==0.19.8.1=hd1a6beb_1008
-  - conda-forge/osx-64::gmp==6.2.1=h2e338ed_0
-  - conda-forge/osx-64::lerc==3.0=he49afe7_0
-  - conda-forge/osx-64::libbrotlidec==1.0.9=h0d85af4_6
-  - conda-forge/osx-64::libbrotlienc==1.0.9=h0d85af4_6
-  - conda-forge/osx-64::libxcb==1.13=h0d85af4_1004
-  - conda-forge/osx-64::lz4-c==1.9.3=he49afe7_1
-  - openssl
-  - conda-forge/osx-64::readline==8.1=h05e3726_0
-  - conda-forge/osx-64::tbb==2021.5.0=h940c156_0
-  - conda-forge/osx-64::tk==8.6.12=h5dbffcc_0
-  - conda-forge/osx-64::zlib==1.2.11=h9173be1_1013
-  - conda-forge/osx-64::brotli-bin==1.0.9=h0d85af4_6
-  - conda-forge/osx-64::gnutls==3.6.13=h756fd2b_1
-  - conda-forge/osx-64::libpng==1.6.37=h7cec526_2
-  - conda-forge/osx-64::mkl==2021.4.0=h89fa619_689
-  - conda-forge/osx-64::openh264==2.1.1=hfd3ada9_0
-  - conda-forge/osx-64::sqlite==3.37.0=h23a322b_0
-  - conda-forge/osx-64::zstd==1.5.2=h582d3a0_0
-  - conda-forge/osx-64::brotli==1.0.9=h0d85af4_6
-  - conda-forge/osx-64::freetype==2.10.4=h4cff582_1
-  - conda-forge/osx-64::libtiff==4.3.0=h17f2ce3_3
-  - conda-forge/osx-64::python==3.9.10=hea1dfa3_2_cpython
-  - conda-forge/noarch::cycler==0.11.0=pyhd8ed1ab_0
-  - pytorch/osx-64::ffmpeg==4.3=h0a44026_0
-  - conda-forge/osx-64::lcms2==2.12=h577c468_0
-  - conda-forge/osx-64::libwebp==1.2.2=h28dabe5_0
-  - conda-forge/noarch::munkres==1.1.4=pyh9f0ad1d_0
-  - conda-forge/osx-64::openjpeg==2.4.0=h6e7aa92_1
-  - conda-forge/noarch::pyparsing==3.0.7=pyhd8ed1ab_0
-  - conda-forge/osx-64::python_abi==3.9=2_cp39
-  - conda-forge/noarch::six==1.16.0=pyh6c4a22f_0
-  - conda-forge/noarch::typing_extensions==4.1.1=pyha770c72_0
-  - conda-forge/noarch::wheel==0.37.1=pyhd8ed1ab_0
-  - certifi
-  - conda-forge/osx-64::kiwisolver==1.3.2=py39hf018cea_1
-  - conda-forge/osx-64::mkl-service==2.4.0=py39hadd8af7_0
-  - conda-forge/noarch::packaging==21.3=pyhd8ed1ab_0
-  - conda-forge/osx-64::pillow==9.0.1=py39hd2c7aa1_2
-  - conda-forge/noarch::python-dateutil==2.8.2=pyhd8ed1ab_0
-  - pytorch/osx-64::pytorch==1.10.2=py3.9_0
-  - conda-forge/osx-64::setuptools==60.9.3=py39h6e9494a_0
-  - conda-forge/osx-64::tornado==6.1=py39h89e85a6_2
-  - conda-forge/osx-64::unicodedata2==14.0.0=py39h89e85a6_0
-  - conda-forge/osx-64::fonttools==4.29.1=py39h89e85a6_0
-  - defaults/osx-64::numpy-base==1.21.2=py39he0bd621_0
-  - conda-forge/noarch::pip==22.0.3=pyhd8ed1ab_0
-  - conda-forge/osx-64::matplotlib==3.5.1=py39h6e9494a_0
-  - conda-forge/osx-64::matplotlib-base==3.5.1=py39hb07454d_0
-  - conda-forge/osx-64::mkl_fft==1.3.1=py39h7ae3660_1
-  - conda-forge/osx-64::mkl_random==1.2.2=py39h4d6be9b_0
-  - defaults/osx-64::numpy==1.21.2=py39h4b4dc7a_0
-  - pytorch/osx-64::torchvision==0.11.3=py39_cpu
+  - torchvision
   - pytorch-lightning
+  - opencv
 prefix: /Users/shion/opt/anaconda3/envs/siamfc
