+++ siamfc/siamfc.py
@@ -1,40 +1,115 @@
 import torch
 import torch.nn as nn
-import torch.functional as F
-from pytorch_lightning import LightningModule
-from .losses import bce_loss
+import torch.nn.functional as F
+import pytorch_lightning as pl
+import numpy as np
 
 
-class SiamFCNet(LightningModule):
-    def __init__(self, embedding_net, output_scale=0.001) -> None:
+class SiamFCNet(pl.LightningModule):
+    def __init__(self, encoder, batch_size, lr, loss, output_scale=0.001, pretrained=False):
+        """Build SiamFC network
+        
+        Parameters
+        ----------
+        encoder : nn.Module
+            Encoding network to embed target and search images
+        
+        batch_size : int
+            Batch size
+        
+        lr : float
+            Learning rate
+            
+        loss : 
+            Loss function
+            
+        output_scale : float, default=0.01
+            Output scaling factor for response maps
+        """
         super().__init__()
-        self.embedding_net = embedding_net
+        self.encoder = encoder
+        self.batch_size = batch_size
+        self.lr = lr
+        self.loss = loss
+        self._init_weights()
+        
         self.output_scale = output_scale
-    
-    def init_weights(self) -> None:
-        """Initialize weights of embedding network"""
-        self.embedding_net.init_weights()
-    
-    def forward(self, z, x):
-        """Forward pass to calculate score map for search and target images."""
-        target_embedded = self.embedding_net(z)
-        search_embedded = self.embedding_net(x)
-        return self._xcorr(target_embedded, search_embedded)    
+        self.total_stride = 8
+        self.r_pos = 16
+        self.r_neg = 0
 
-    def training_step(self, batch, batch_nb):
-        """Calculates loss for one step in training."""
-        x, y = batch
-        loss = bce_loss(self(x), y)
+    def forward(self, z, x):
+        """Calculate response map for pair of exmplar and search images.
+        
+        Parameters
+        ----------
+        z : array of shape ()
+            Exemplar image
+            
+        x : array of shape ()
+            Search image
+            
+        Returns
+        -------
+        response_map : array of shape ()
+            Cross-correlation response map of embedded images
+        """
+        target_embedded = self.encoder(z)
+        search_embedded = self.encoder(x)
+        return self._xcorr(target_embedded, search_embedded) * self.output_scale    
+
+    def training_step(self, batch, batch_idx): #What does batch_nb mean?
+        """Returns and logs loss for training step."""
+        loss = self._shared_step(batch, batch_idx)
+        # result = pl.TrainResult(minimize=loss, on_epoch=True)
+        # result.log('train_loss', loss, on_epoch=True)
+        return {"loss": loss}
+
+    def validation_step(self, batch, batch_idx):
+        """Returns loss for validation step."""
+        loss = self._shared_step(batch, batch_idx)
+        # result = pl.EvalResult(checkpoint_on=loss)
+        # result.log('avg_val_loss', loss)
+        return {"val_loss": loss}
+
+    def configure_optimizers(self): 
+        """Returns optimizer for model."""
+        optimizer = torch.optim.Adam(self.encoder.parameters(), lr=self.lr)
+        return optimizer
+
+    def _init_weights(self) -> None:
+        """Initialize weights of encoder network."""
+        for m in self.encoder.modules():
+            if isinstance(m, nn.Conv2d):
+                nn.init.xavier_uniform_(m.weight, 1)
+                if m.bias is not None:
+                    nn.init.constant_(m.bias, 0)
+            elif isinstance(m, nn.BatchNorm2d):
+                nn.init.constant_(m.weight, 1)
+                nn.init.constant_(m.bias, 0)
+            elif isinstance(m, nn.Linear):
+                nn.init.xavier_uniform_(m.weight, 1)
+                if m.bias is not None:
+                    nn.init.constant_(m.bias, 0)
+
+    def _shared_step(self, batch, batch_idx):
+        """Returns loss for pass through model with provided batch."""
+        (z, x) = batch
+        
+        # Encode target and search images
+        hz = self.encoder(z)
+        hx = self.encoder(x)
+        
+        #print(hz.shape, hx.shape)
+        
+        # Calculate response map and loss
+        responses = self._xcorr(hz, hx)
+        labels = self._create_labels(responses.size())
+        loss = self.loss(responses, labels)
         return loss
 
-    def configure_optimizer(self):
-        """Returns optimizer for model."""
-        lr = self.hparams.lr
-        opt = torch.optim.Adam(self.embedding_net.parameters(), lr=lr)
-        return opt
-
-    def _xcorr(z, x):
-        """Calculates cross-correlation between target and search embeddings.
+    def _xcorr(self, z, x):
+        """Calculates cross-correlation between target and search image embeddings.
         
         Parameters
         ----------
@@ -51,3 +126,37 @@ class SiamFCNet(LightningModule):
         """
         score_map = F.conv2d(x, z)
         return score_map
+    
+    def _create_labels(self, size):
+        # skip if same sized labels already created
+        if hasattr(self, 'labels') and self.labels.size() == size:
+            return self.labels
+
+        def logistic_labels(x, y, r_pos, r_neg):
+            dist = np.abs(x) + np.abs(y)  # block distance
+            labels = np.where(dist <= r_pos,
+                              np.ones_like(x),
+                              np.where(dist < r_neg,
+                                       np.ones_like(x) * 0.5,
+                                       np.zeros_like(x)))
+            return labels
+
+        # distances along x- and y-axis
+        n, c, h, w = size
+        x = np.arange(w) - (w - 1) / 2
+        y = np.arange(h) - (h - 1) / 2
+        x, y = np.meshgrid(x, y)
+
+        # create logistic labels
+        r_pos = self.r_pos / self.total_stride
+        r_neg = self.r_neg / self.total_stride
+        labels = logistic_labels(x, y, r_pos, r_neg)
+
+        # repeat to size
+        labels = labels.reshape((1, 1, h, w))
+        labels = np.tile(labels, (n, c, 1, 1))
+
+        # convert to tensors
+        self.labels = torch.from_numpy(labels).to(self.device).float()
+        
+        return self.labels
